requirements.txt
streamlit
scikit-learn==1.5.1
pandas
numpy
feature-engine
matplotlib
seaborn
plotly
pillow